Add support for setting ACL rules on windows certificate store keys
Added support via the win_acl module which now supports the Cert:\<store location>\<storename>\<thumbprint> paths.
The code for this is only available through the C api for CAPI (crypto api) and CNG (Crypto API: Next Generation).
The interface to CAPI and CNG is implemented as a module util called CertACLHelper.