win_updates - Improve file locking behaviour
Improve the file locking behaviour to allow the polling task to safely
open a handle on the output status file. This changes the failure
condition to log the failure message with a FILE_SHARE_READ handle
allowing others to open a read handle to the same file at the same time.