win_updates - Improve batch runner (#495)
Improve the batch runner bootstrapper that is used in win_updates to run
the update code inside a batch process. This is done by starting an
ephemeral process through the task scheduler that is unrelated to
PowerShell and using a new API to spawn a new PowerShell process as a
child of that ephemeral task. Unlike the ephemeral task process, the
newly spawned process exposes the stdio handles allowing the module to
have better visibility on failures and to send information to run than
before.