win_updates - Improve locking handling
Add a retry mechanism to the output polling code to retry opening the
result file if it failed due to a lock. While it is unknown what causes
the lock this should at least provide a better mechanism to retry the
operation in case it was another process like an antivirus that
temporarily locked the file.