win_updates - fix reboot loop behaviour (#542)
Fixes up the logic in the action plugin to not retry if the same update
was found to be installed. This should pickup scenarios where an update
was rolled back during a reboot or WUA failed to detect the install
failed to install.

Also adds extra handling for a failure of KB2267602 which contains
Microsoft Defender definitions to try and install a failed update using
the Defender executable. This has shown to be more reliable than WUA
when it comes to installing this update.