Add tls13 support for win_get_url (#573)
* Add tls13 support

Using the win_get_url module I was getting the following error: 
"The underlying connection was closed: Could not establish trust relationship for the SSL/TLS secure channel"

However, from the machine the download was working fine with in a powershell terminal. After some troubleshooting, I noticed that the issue was with TLS1.3.

Adding an extra condition for TLS1.3 actually fixed the issue. Please let me know what you think.

* Checking net framework version instead of explicitly enabling tls1.3

* Fix syntax errors

* Fix trailing whitespace

* fix whitespace trailing

* Add changelog fragment

* Update plugins/module_utils/WebRequest.psm1

Co-authored-by: Jordan Borean <jborean93@gmail.com>

---------

Co-authored-by: yacine-learning <yacine-learning@github.com>
Co-authored-by: Jordan Borean <jborean93@gmail.com>