setup - remvoe dep on shared function
Remove the dep on ConvertFrom-AnsibleJson so that we can start to
deprecate that function in Ansible exec wrapper. This is the only
instance where it is used outside of the exec wrapper and it can easily
be replaced by the custom PSCustomObject enumerator. Luckily this code
path is not used by any of our supported connection plugins. It may be
used by other connection types that cannot pipeline or through local
debugging.